Agrego pantalla de simulacion con ETA

+++ app/src/main/java/com/jk/rcp/main/utils/Constants.java
@@ -6,5 +6,5 @@ public class Constants {
     public static final Double APRENDIZAJE_DURACION_SEGUNDOS_COMPRESION_INSUFLACION = 30.0;
     public static final Double APRENDIZAJE_DURACION_SEGUNDOS_FINAL = 68.0;
     public static final Double SIMULACION_DURACION_SEGUNDOS_PASO2 = 10.0;
-    public static final Double SIMULACION_DURACION_SEGUNDOS_PASO3 = 60.0;
+    public static final int SIMULACION_DURACION_SEGUNDOS_PASO3 = 60;
 }
